feat(cni-installer): rename pkg -> internal
