+++ internal/cni/conf_test.go
