+++ internal/cni/installer_test.go
