+++ internal/cni/remove_test.go
