+++ internal/cni/testdata/auth-token
