+++ internal/cni/testdata/auth-token-new
