+++ internal/cni/testdata/bin/cni-binary
