+++ internal/cni/testdata/cni-binary
