+++ internal/cni/testdata/cni-invalid-plugin.json
