+++ internal/cni/testdata/cni-no-plugins.json
