+++ internal/cni/testdata/cni-src.json
