+++ internal/cni/testdata/kubeconfig-exp-env.yaml
