+++ internal/cni/testdata/kubeconfig-exp-sibling.yaml
