+++ internal/cni/testdata/kubeconfig-src.yaml
